In this version, while the cserver and kserver add user to the map, I also store the corresponding mapping in two files, respectively, each piece of information is stored in one line in the file. While the cserver and kserver are started, they read the registered user information from the files. Thus, if a user Alice registered, then even if both of the two servers are down, while they restart, then can retrieve the registered user information immediately.
The test on cserver is passed and completed, it's stable. However, at kserver, public key of encrypted private key of a user is always read as "null", it's very weird, I guess that's because these two keys are very long, they caused some error while writing or reading the file. Please look at it if you have time.

By the way, we still need to generate passwords for other teams and TA.